move theory into its own module
